Nav now hides on hp; changed Tabs -> Navigation; added Prettier rule
