+++ .prettierrc.json
@@ -2,5 +2,6 @@
   "trailingComma": "es5",
   "tabWidth": 2,
   "semi": true,
-  "singleQuote": false
+  "singleQuote": false,
+  "html.format.wrapAttributes": "force-aligned"
 }
