+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import TabsBar from "../components/TabsBar";
+import Navigation from "../components/Navigation";
 import GiantHeader from "../components/GiantHeader";
 import type { AppProps } from "next/app";
 import "../styles/globals.css";
@@ -7,12 +7,12 @@ function MyApp({ Component, pageProps }: AppProps) {
   return (
     <>
       <div className="h-screen w-screen bg-gray-900">
-        <TabsBar>
+        <Navigation>
           <div className="pt-6 lg:pt-12 pl-28 lg:pl-44 pr-4">
             <Component {...pageProps} />
           </div>
           <GiantHeader />
-        </TabsBar>
+        </Navigation>
       </div>
     </>
   );
